Bike Sharing Demand
The main objective of this study is to investigate how environmental and seasonal factors affect the number of bicycle rentals per day. Furthermore, the model was improved to make future predictions about daily bicycle rental used.
The study will ultimately be beneficial for businesses in achieving the demand for sharing bikes by maintaining the supply.